Added the CATCH_REQUIRE_LONG_STRING() macro and catch_compare_long_strings() function.
This greatly enhance the speed at which I can fix such long strings.
The default Catch2 just writes them on the screen one after the other
but it does not show the differences. For short strings that's okay,
but fr long ones, it's close to useless. With the compare_long_strings()
function, we print differences (a bit like a diff) on the screen showing
each character that differs. Since I have that I was able to fix problems
in no time.